implemented authentication for users

+++ frontend/src/components/WorkoutDetails/WorkoutDetails.jsx
@@ -2,14 +2,25 @@ import React from "react";
 import { useWorkoutsContext } from "../../hooks/useWorkoutContext";
 import DeleteIcon from "../../assets/delete.png"
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
+import { useAuthContext } from "../../hooks/useAuthContext";
 
 
 const WorkoutDetails = ({ workout }) => {
 const {dispatch} = useWorkoutsContext()
+const {user} = useAuthContext();
 
   const handleClick = async () =>{
+
+    if (!user) {
+      setError("You must be logged in");
+      return;
+    }
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
-      method:"DELETE"
+      method:"DELETE",
+      headers:{
+        "Authorization":`Bearer ${user.token}`
+      }
     })
     const json = await response.json()
